Refactor: create src folder and change file names
