+++ src/includingHyphenWordExtractor.js
