+++ src/styledComponentsExtractor.js
